medical_communication_app/requirements.txt
fastapi
uvicorn
pydantic
jinja2
python-multipart
pydub
whisper
groq
torch
transformers
sentencepiece
scikit-learn
numpy
PyYAML
pyannote.audio